Handle sets properly
Since we replace Java objects with a placeholder, we might mess up sets. This
can happen in at least two cases:

- if we get the same string representation for two items in a set, they would be
  collapsed into one item in our resulting set

- if we use a sorted set, the comparator function probably wouldn't work anymore
  afterwards, making for all sorts of shenanigans.

This brings us to this unfortunate hack: We now send sets to the client as
vectors, with an extra :prone.prep/set? item added to the end. This is to keep
the same nesting, while efficiently being able to tell if a vector represents a
vector or a set. Oh well. Oh well indeed.